add UserBook UserBookDAO and refactor Book

+++ src/main/java/com/github/magdanadratowska/model/Book.java
@@ -2,12 +2,12 @@ package com.github.magdanadratowska.model;
 
 public class Book {
 
-    protected int id;
+    protected long id;
     protected String title;
     protected String authorName;
     protected String authorSurname;
 
-    public Book(int id, String title, String authorName, String authorSurname) {
+    public Book(long id, String title, String authorName, String authorSurname) {
         this.id = id;
         this.title = title;
         this.authorName = authorName;
@@ -23,11 +23,11 @@ public class Book {
     public Book() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
